Create onvif-url.xml
Example metadata about the ONVIF device_service endpoint the camera seems to try and send up in one of its payloads to the remote servers. 

Gives us a few clues on what may be possible with ONVIF as the device_service endpoint can be used for control including PTZ but at the moment the actual web server running on 80 does not return anything when you try and call the service locally.